requirements.txt
Django==4.2.23
djangorestframework==3.14.0
django-filter==23.5
beautifulsoup4==4.12.2
requests==2.31.0
django-cors-headers==4.3.1
django-extensions==3.2.3
ipython==8.17.2